fix(syncexpirytime): to typescript and use the new modules
