+++ services/viva-ms/tsconfig.json
@@ -3,7 +3,8 @@
   "compilerOptions": {
     "outDir": "dist",
     "baseUrl": "src",
-    "rootDir": "./src"
+    "rootDir": "./src",
+    "noImplicitAny": false,
   },
   "exclude": [],
   "include": ["src/**/*"]
